Initial Refactor * App.java -> Endpoints.java * Module.java -> GennyModule.java * Created basic Package Structure * Started removing Ask/EntityEntity/AttributeLink stuff * Started cleaning up excess code
