+++ bootq/src/main/java/life/genny/bootq/models/SheetState.java
@@ -3,6 +3,8 @@ package life.genny.bootq.models;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
+import life.genny.bootq.sheets.RealmUnit;
+
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
