+++ bootq/src/main/java/life/genny/bootq/sheets/DataUnit.java
@@ -1,23 +1,26 @@
-package life.genny.bootq.models;
+package life.genny.bootq.sheets;
 
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * A POJO representation of a standard "Data Sheet" containing all the raw data needed to 
+ * bootstrap into a project
+ */
 public abstract class DataUnit {
 
     public DataUnit() {
     }
 
+    protected String name;
+
+    // TODO: Reduce the complexity of these Maps by making better use of POJOs
     protected Map<String, Map<String, String>> attributes = new HashMap<>();
-    protected Map<String, Map<String, String>> attributeLinks = new HashMap<>();
     protected Map<String, Map<String, String>> questionQuestions = new HashMap<>();
     protected Map<String, Map<String, String>> validations = new HashMap<>();
     protected Map<String, Map<String, String>> dataTypes = new HashMap<>();
     protected Map<String, Map<String, String>> questions = new HashMap<>();
-    protected Map<String, Map<String, String>> asks = new HashMap<>();
-    protected Map<String, Map<String, String>> notifications = new HashMap<>();
     protected Map<String, Map<String, String>> entityAttributes = new HashMap<>();
-    protected Map<String, Map<String, String>> entityEntitys = new HashMap<>();
     protected Map<String, Map<String, String>> baseEntitys = new HashMap<>();
     protected Map<String, Map<String, String>> messages = new HashMap<>();
     protected Map<String, Map<String, String>> def_baseEntitys = new HashMap<>();
@@ -26,11 +29,7 @@ public abstract class DataUnit {
     public void setAttributes(Map<String, Map<String, String>> attributes) {
         this.attributes = attributes;
     }
-
-    public void setAttributeLinks(Map<String, Map<String, String>> attributeLinks) {
-        this.attributeLinks = attributeLinks;
-    }
-
+    
     public void setQuestionQuestions(Map<String, Map<String, String>> questionQuestions) {
         this.questionQuestions = questionQuestions;
     }
@@ -47,20 +46,8 @@ public abstract class DataUnit {
         this.questions = questions;
     }
 
-    public void setAsks(Map<String, Map<String, String>> asks) {
-        this.asks = asks;
-    }
-
-    public void setNotifications(Map<String, Map<String, String>> notifications) {
-        this.notifications = notifications;
-    }
-
     public void setEntityAttributes(Map<String, Map<String, String>> entityAttributes) {
         this.entityAttributes = entityAttributes;
-    }
-
-    public void setEntityEntitys(Map<String, Map<String, String>> entityEntitys) {
-        this.entityEntitys = entityEntitys;
     }
 
     public void setBaseEntitys(Map<String, Map<String, String>> baseEntitys) {
@@ -71,13 +58,8 @@ public abstract class DataUnit {
         this.messages = messages;
     }
 
-
     public Map<String, Map<String, String>> getAttributes() {
         return attributes;
-    }
-
-    public Map<String, Map<String, String>> getAttributeLinks() {
-        return attributeLinks;
     }
 
     public Map<String, Map<String, String>> getQuestionQuestions() {
@@ -96,20 +78,8 @@ public abstract class DataUnit {
         return questions;
     }
 
-    public Map<String, Map<String, String>> getAsks() {
-        return asks;
-    }
-
-    public Map<String, Map<String, String>> getNotifications() {
-        return notifications;
-    }
-
     public Map<String, Map<String, String>> getEntityAttributes() {
         return entityAttributes;
-    }
-
-    public Map<String, Map<String, String>> getEntityEntitys() {
-        return entityEntitys;
     }
 
     public Map<String, Map<String, String>> getBaseEntitys() {
@@ -127,5 +97,14 @@ public abstract class DataUnit {
     public Map<String, Map<String, String>> getMessages() {
         return messages;
     }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
 
 }
